Restructured Enums
Made individual enum classes
