+++ TrainModel/src/trainmodel/model/TrainModelInterface.java
@@ -1,8 +1,8 @@
 package trainmodel.model;
 
-import utils.train.TrainModelEnums.DoorStatus;
-import utils.train.TrainModelEnums.OnOffStatus;
-import utils.train.TrainModelEnums.Failure;
+import utils.train.DoorStatus;
+import utils.train.OnOffStatus;
+import utils.train.Failure;
 
 
 /**
